Revert "refactor: :recycle: modify `term_list` struct"
